requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
dataclasses-json==0.6.7
distro==1.9.0
filetype==1.2.0
frozenlist==1.8.0
google-auth==2.47.0
google-genai==1.59.0
google-generativeai
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.6
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-experimental==0.4.1
langchain-google-genai==4.2.0
langchain-text-splitters==1.1.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
marshmallow==3.26.2
matplotlib==3.10.8
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.4.1
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.45
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.13.0
uv==0.9.26
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.27.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
PyMuPDF>=1.23.0
camelot-py[cv]>=0.11.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pocket-tts
compressed-tensors
faster-whisper
numpy
websockets
soundfile
fastapi
routers
python-dotenv

-e .